Refactoring and rebase
