+++ client/TEMT6000_sensor.py
@@ -6,9 +6,9 @@ from adafruit_ads1x15.ads1115 import ADS1115
 from adafruit_ads1x15.ads1115 import P0
 from adafruit_ads1x15.analog_in import AnalogIn
 from threading import Thread
-import servo_motor
+import client.servo_motor as servo_motor
 import queue
-import tcp_connection
+import client.tcp_connection as tcp_connection
 
 pi = pigpio.pi()
 LOW = 1
