+++ client/main.py
@@ -1,11 +1,11 @@
 import time
 import pigpio
-import tcp_connection
-import principal_thread
-import mode_controller
+import client.tcp_connection as tcp_connection
+import client.principal_thread as principal_thread
+import client.mode_controller as mode_controller
 from threading import Thread
 import queue
-import servo_motor
+import client.servo_motor as servo_motor
 
 data_queue = queue.Queue()
 stop_queue = queue.Queue()
