+++ client/mode_controller.py
@@ -1,7 +1,7 @@
-import tcp_connection
-import principal_thread
-import jinglebell
-import TEMT6000_sensor
+import client.tcp_connection as tcp_connection
+import client.principal_thread as principal_thread
+import client.jinglebell as jinglebell
+import client.TEMT6000_sensor as TEMT6000_sensor
 from threading import Thread
 
 # Les print en commentaire c'était pour les test si vous souhaitez voir dans le terminal qu'est-ce que le script envoie
